Added main_website and taxonomy components
Main website is for serving the volunteer front-end. This uses
app.js, css/js, and pug views currently located in the web
component. It is to create a far simpler, focused component
as the entry point for devs.

Taxonomy is for defining the taxonomy data for each brigade.
In the future this will make use of TOML or other format.

All code in the web component used for Project Match will be
moved out of web, including the REST API.